[Config] Change user country and language cvars to strings
Replaced the int version of user country and language with strings
for easier user interaction in the config dialog which now presents
a proper dropdown of values. Added some logic for config parser to
fall through to default values for type mismatches to allow startup
with old versions of the config.